chore: refactor widgets structure
